# lib/models/userModel.ts
import * as mongoose from "mongoose";

const Schema = mongoose.Schema;

export const UserSchema : mongoose.Schema = new Schema({
    name: {
        type: String,
        required: [true, "a name is required"],
        min: [1, "a name requires minimum 1 letter"] 
    },
    created: String
});

export interface IMongoUser extends IUser, mongoose.Document{}
export interface IUser {
    name: string,
    created: Date
}